+++ CSSP/src/main/java/com/soshow/ssi/activity/service/impl/ActivityServiceImpl.java
@@ -42,7 +42,7 @@ public class ActivityServiceImpl implements ActivityService{
 		return activityDao.findPageByCondition(pc);
 	}
 	
-	public Integer countByCondition(ActivityCondition pc) {
+	public Long countByCondition(ActivityCondition pc) {
 		return activityDao.countByCondition(pc);
 	}
 }
