+++ CSSP/src/main/java/com/soshow/ssi/clubActivity/controller/ClubActivityController.java
@@ -122,7 +122,7 @@ public class ClubActivityController extends BaseController{
 			if(condition.getPageSize()>0){
 				condition.init();
 			}
-			int count =  clubActivityService.countByCondition(condition);
+			long count =  clubActivityService.countByCondition(condition);
 			if(count==0){
 				response.setStatusResponse(CommStatusEnum.NOFIND);
 				return response;
