+++ CSSP/src/main/java/com/soshow/ssi/clubActivity/service/ClubActivityService.java
@@ -38,6 +38,6 @@ public interface ClubActivityService {
 	/**
 	 * 统计查询数据的条数
 	 */
-	public Integer countByCondition(ClubActivityCondition pc);
+	public Long countByCondition(ClubActivityCondition pc);
 	
 }
