+++ CSSP/src/main/java/com/soshow/ssi/company/controller/CompanyController.java
@@ -122,7 +122,7 @@ public class CompanyController extends BaseController{
 			if(condition.getPageSize()>0){
 				condition.init();
 			}
-			int count =  companyService.countByCondition(condition);
+			long count =  companyService.countByCondition(condition);
 			if(count==0){
 				response.setStatusResponse(CommStatusEnum.NOFIND);
 				return response;
