+++ CSSP/src/main/java/com/soshow/ssi/experience/controller/ExperienceController.java
@@ -122,7 +122,7 @@ public class ExperienceController extends BaseController{
 			if(condition.getPageSize()>0){
 				condition.init();
 			}
-			int count =  experienceService.countByCondition(condition);
+			long count =  experienceService.countByCondition(condition);
 			if(count==0){
 				response.setStatusResponse(CommStatusEnum.NOFIND);
 				return response;
