+++ CSSP/src/main/java/com/soshow/ssi/goodsOrderDetailed/controller/GoodsOrderDetailedController.java
@@ -122,7 +122,7 @@ public class GoodsOrderDetailedController extends BaseController{
 			if(condition.getPageSize()>0){
 				condition.init();
 			}
-			int count =  goodsOrderDetailedService.countByCondition(condition);
+			long count =  goodsOrderDetailedService.countByCondition(condition);
 			if(count==0){
 				response.setStatusResponse(CommStatusEnum.NOFIND);
 				return response;
