+++ CSSP/src/main/java/com/soshow/ssi/notice/service/NoticeService.java
@@ -38,6 +38,6 @@ public interface NoticeService {
 	/**
 	 * 统计查询数据的条数
 	 */
-	public Integer countByCondition(NoticeCondition pc);
+	public Long countByCondition(NoticeCondition pc);
 	
 }
