+++ CSSP/src/main/java/com/soshow/ssi/resume/controller/ResumeController.java
@@ -122,7 +122,7 @@ public class ResumeController extends BaseController{
 			if(condition.getPageSize()>0){
 				condition.init();
 			}
-			int count =  resumeService.countByCondition(condition);
+			long count =  resumeService.countByCondition(condition);
 			if(count==0){
 				response.setStatusResponse(CommStatusEnum.NOFIND);
 				return response;
