+++ CSSP/src/main/java/com/soshow/ssi/student/controller/StudentController.java
@@ -173,7 +173,7 @@ public class StudentController extends BaseController{
 			if(condition.getPageSize()>0){
 				condition.init();
 			}
-			int count =  studentService.countByCondition(condition);
+			long count =  studentService.countByCondition(condition);
 			if(count==0){
 				response.setStatusResponse(CommStatusEnum.NOFIND);
 				return response;
